Fix Shadow Issues With Point Lights

+++ Source/Core/Renderer/ERS_CLASS_VisualRenderer/ERS_CLASS_ShadowMaps.cpp
@@ -60,7 +60,7 @@ void ERS_CLASS_ShadowMaps::UpdateShadowMaps(ERS_STRUCT_Shader* DepthMapShader, E
 
         }
     } else if (UpdateMode != ERS::Renderer::ERS_SHADOW_UPDATE_MODE_CONSECUTIVE) {
-        for (unsigned int i = 0; i < SystemUtils_->RendererSettings_->MaxShadowUpdatesPerFrame_; i++) {
+        for (unsigned int i = 0; i < (unsigned int)SystemUtils_->RendererSettings_->MaxShadowUpdatesPerFrame_; i++) {
 
             // Calculate The Current index, Wrap At End Of List Size
             LastUpdateIndex_++;
